pyproject.toml
[build-system]
requires = [ "setuptools>=41", "wheel", ]
build-backend = "setuptools.build_meta"

[project]
name = "pyrite_framework"
version = "0.12.1"
authors = [
    { name="BetterBuiltFool", email="betterbuiltfool@gmail.com"}
]
description = "A simple yet featureful framework for simplifying pygame game creation."
readme = "README.md"
requires-python = ">=3.12.4"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Games/Entertainment",
    "Topic :: Software Development :: Libraries :: pygame",
    "Typing :: Typed"
]
dependencies = [
    "pygame_ce >= 2.5.2",
    "pyglm >= 2.8.3",
    "weaktree >= 0.1.0, < 1"
]